+++ CHANGELOG.md
@@ -4,6 +4,10 @@ FEATURES
 * modules/acl-controller: Add `assign_public_ip` variable to the ACL controller
   to support running on public subnets.
   [[GH-64](https://github.com/hashicorp/terraform-aws-consul-ecs/pull/64)]
+* modules/mesh-task: Add `application_shutdown_delay_seconds` variable to
+  delay application shutdown. This allows time for incoming traffic to drain
+  off for better graceful shutdown.
+  [[GH-67](https://github.com/hashicorp/terraform-aws-consul-ecs/pull/67)]
 
 ## 0.2.0 (Nov 16, 2021)